# crypto: faster keccak on Address and Hash