--- a/crypto/fixed/keccakf.go
+++ b/crypto/fixed/keccakf.go
@@ -1,10 +1,9 @@
 // This file is copied from golang/x/crypto/sha3/keccakf.go
-
 // Copyright 2014 The Go Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-//  +build !amd64 appengine gccgo
+//  +build !amd64 purego !gc
 
 package fixed