--- a/crypto/fixed/keccakf_amd64.go
+++ b/crypto/fixed/keccakf_amd64.go
@@ -4,7 +4,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// +build amd64,!appengine,!gccgo
+// +build amd64,!purego,gc
 
 package fixed