--- a/crypto/fixed/keccakf_amd64.go
+++ b/crypto/fixed/keccakf_amd64.go
@@ -12,4 +12,4 @@ package fixed
 
 //go:noescape
 
-func keccakF1600(a *[25]uint64)
+func keccakF1600(state *[25]uint64)