… trie/secure_trie.go