This is a bit of an experiment, which turned out pretty nicely. 
This PR makes it possible to do keccak hashes of `[20]byte` and `[32]byte` (i.e: `Address` and `Hash`) 2-3x faster and without any allocs:
```
[user@work crypto]$ go test . -run - -bench KeccakHash
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/crypto
BenchmarkKeccakHash/generic-20-6                 1363515               976 ns/op             480 B/op          2 allocs/op
BenchmarkKeccakHash/fixed-20-6                   2680185               473 ns/op               0 B/op          0 allocs/op
BenchmarkKeccakHash/generic-32-6                  808141              1263 ns/op             480 B/op          2 allocs/op
BenchmarkKeccakHash/fixed-32-6                   2611064               448 ns/op               0 B/op          0 allocs/op
```
So far, this PR isnt' actually used anywhere, but we have a few places in the code where we do these operations. I'll integrate it into those places and convert the draft into an actual PR.