@holiman 2020-12-15
Funky, `govet` complains about the assembly I imported from upstream

![Screenshot_2020-12-15 Job #24450 1 - ethereum go-ethereum - Travis CI](https://user-images.githubusercontent.com/142290/102200928-c0429300-3ec5-11eb-8174-35460a50b329.png)